Validation fixes on GET/ DELETE/

+++ controller/userController.js
@@ -4,7 +4,6 @@ const upload = require("../config/multer");
 const statsd = require("../config/statsD");
 const userService = require("../service/userService");
 
-
 const validateUserFields = (userData, requiredfields, next, allowedField) => {
   if (requiredfields) {
     for (const field of requiredfields) {
@@ -65,7 +64,6 @@ const validateUserFields = (userData, requiredfields, next, allowedField) => {
 };
 
 const createUser = (req, res, next) => {
-
   const userData = req.body;
 
   const requiredfields = new Set([
@@ -86,8 +84,16 @@ const createUser = (req, res, next) => {
   userService
     .createUser(userData, `${req.method}_${req.originalUrl}`)
     .then((response) => {
-   
-      logData(req.method, req.originalUrl,  req.get('user-agent'), 'info', req.body, 200, 'User created successfully', response);
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        200,
+        "User created successfully",
+        response
+      );
 
       return res.status(201).json({ message: "User Created", data: response });
     })
@@ -97,27 +103,35 @@ const createUser = (req, res, next) => {
 };
 
 const getAUser = (req, res, next) => {
- // Check if there is a request body
- if (req.body && Object.keys(req.body).length > 0) {
-  const error = new Error("Request body not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
+  // Check if there is a request body
+  if (req.body && Object.keys(req.body).length > 0) {
+    const error = new Error("Request body not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
 
-// Check if Content-Type is multipart/form-data
-const contentType = req.headers['content-type'];
-if (contentType && contentType.startsWith('multipart/form-data')) {
-  const error = new Error("Multipart form data not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
-const email = req.authenticatedUser;
+  // Check if Content-Type is multipart/form-data
+  const contentType = req.headers["content-type"];
+  if (contentType && contentType.startsWith("multipart/form-data")) {
+    const error = new Error("Multipart form data not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
+  const email = req.authenticatedUser;
 
   userService
     .getAUser(email, `${req.method}_${req.originalUrl}`)
     .then((user) => {
-
-      logData(req.method, req.originalUrl,  req.get('user-agent'), 'info', req.body, 200, 'User fetched successfully', user);
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        200,
+        "User fetched successfully",
+        user
+      );
 
       return res.status(200).json(user);
     })
@@ -135,8 +149,16 @@ const updateUser = (req, res, next) => {
   userService
     .updateUser(userId, userData, `${req.method}_${req.originalUrl}`)
     .then((updatedUser) => {
-  
-      logData(req.method, req.originalUrl, req.get('user-agent'), 'info', req.body, 200,'User updated successfully', updatedUser);
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        200,
+        "User updated successfully",
+        updatedUser
+      );
 
       return res.status(204).end();
     })
@@ -145,76 +167,83 @@ const updateUser = (req, res, next) => {
     });
 };
 
-
 const uploadProfilePic = (req, res, next) => {
- // Check if there is a request body
- if (req.body && Object.keys(req.body).length > 0) {
-  const error = new Error("Request body not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
+  // Check if there is a request body
+  if (req.body && Object.keys(req.body).length > 0) {
+    const error = new Error("Request body not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
 
-// Check if Content-Type is multipart/form-data
-const contentType = req.headers['content-type'];
-if (contentType && contentType.startsWith('multipart/form-data')) {
-  const error = new Error("Multipart form data not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
-
-  const file = req.file
+  const file = req.file;
   const userEmail = req.authenticatedUser;
-  const allowedImageTypes = ['image/png', 'image/jpg', 'image/jpeg'];
+  const allowedImageTypes = ["image/png", "image/jpg", "image/jpeg"];
 
-    if (!file) 
-    {
-      const error = new Error("File not received");
-      error.statusCode = 400;
-      return next(error);
-    }
-    if (!allowedImageTypes.includes(file.mimetype)) {
+  if (!file) {
+    const error = new Error("File not received");
+    error.statusCode = 400;
+    return next(error);
+  }
+  if (!allowedImageTypes.includes(file.mimetype)) {
+    const error = new Error(
+      "Invalid file type. Only PNG, JPG, and JPEG are allowed!"
+    );
+    error.statusCode = 400;
+    return next(error);
+  }
 
-      const error = new Error("Invalid file type. Only PNG, JPG, and JPEG are allowed!");
-      error.statusCode = 400;
-      return next(error);
-    }
-
-    userService
+  userService
     .uploadProfilePic(userEmail, file, `${req.method}_${req.originalUrl}`)
     .then((data) => {
-      logData(req.method, req.originalUrl, req.get('user-agent'), 'info', req.body, 201,'Profile pic uploaded successfully', data);
-      
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        201,
+        "Profile pic uploaded successfully",
+        data
+      );
+
       return res.status(201).json(data);
     })
     .catch((err) => {
       return next(err);
     });
-
-}
-
+};
 
 const getProfilePic = (req, res, next) => {
- // Check if there is a request body
- if (req.body && Object.keys(req.body).length > 0) {
-  const error = new Error("Request body not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
+  // Check if there is a request body
+  if (req.body && Object.keys(req.body).length > 0) {
+    const error = new Error("Request body not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
 
-// Check if Content-Type is multipart/form-data
-const contentType = req.headers['content-type'];
-if (contentType && contentType.startsWith('multipart/form-data')) {
-  const error = new Error("Multipart form data not allowed in GET requests");
-  error.statusCode = 400;
-  return next(error);
-}
-const userEmail = req.authenticatedUser; 
+  // Check if Content-Type is multipart/form-data
+  const contentType = req.headers["content-type"];
+  if (contentType && contentType.startsWith("multipart/form-data")) {
+    const error = new Error("Multipart form data not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
+  const userEmail = req.authenticatedUser;
 
   userService
     .getProfilePic(userEmail, `${req.method}_${req.originalUrl}`)
     .then((data) => {
-      logData(req.method, req.originalUrl, req.get('user-agent'), 'info', req.body, 200, 'Profile picture retrieved successfully', data);
-      
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        200,
+        "Profile picture retrieved successfully",
+        data
+      );
+
       return res.status(200).json(data);
     })
     .catch((err) => {
@@ -228,12 +257,27 @@ const deleteProfilePic = (req, res, next) => {
     error.statusCode = 400;
     return next(error);
   }
-  const userEmail = req.authenticatedUser; 
+  // Check if Content-Type is multipart/form-data
+  const contentType = req.headers["content-type"];
+  if (contentType && contentType.startsWith("multipart/form-data")) {
+    const error = new Error("Multipart form data not allowed in GET requests");
+    error.statusCode = 400;
+    return next(error);
+  }
+  const userEmail = req.authenticatedUser;
 
   userService
     .deleteProfilePic(userEmail, `${req.method}_${req.originalUrl}`)
     .then(() => {
-      logData(req.method, req.originalUrl, req.get('user-agent'), 'info', req.body, 204, 'Profile picture deleted successfully');
+      logData(
+        req.method,
+        req.originalUrl,
+        req.get("user-agent"),
+        "info",
+        req.body,
+        204,
+        "Profile picture deleted successfully"
+      );
       return res.status(204).end();
     })
     .catch((err) => {
@@ -248,5 +292,5 @@ module.exports = {
   validateUserFields,
   uploadProfilePic,
   getProfilePic,
-  deleteProfilePic
+  deleteProfilePic,
 };
